Fix imports and modularize `src/`
This is a follow-up to PR #43:
* turns `src/` into module
* eliminate any use of the `sys.path.insert(0, 'src')` hack
* use py3-compatible syntax for relative imports
* doesn't have PR #43's conflicts
* IDE's like PyCharm should no longer complain

Outstanding issues not handled by this PR:
* there's several imports which are never used and should be cleaned out
* there's a few variables which aren't defined in their context, e.g. `X`
