+++ style.py
@@ -1,10 +1,8 @@
 from __future__ import print_function
-import sys, os, pdb
-sys.path.insert(0, 'src')
-import numpy as np, scipy.misc 
-from optimize import optimize
+import os, pdb
+from src.optimize import optimize
 from argparse import ArgumentParser
-from utils import save_img, get_img, exists, list_files
+from src.utils import save_img, get_img, exists, list_files
 import evaluate
 
 CONTENT_WEIGHT = 7.5e0
