+++ transform_video.py
@@ -1,9 +1,7 @@
 from __future__ import print_function
 from argparse import ArgumentParser
-import sys
-sys.path.insert(0, 'src')
 import os, random, subprocess, evaluate, shutil
-from utils import exists, list_files
+from src.utils import exists, list_files
 import pdb
 
 TMP_DIR = '.fns_frames_%s/' % random.randint(0,99999)
